changed tests structure for fields subfields
